+++ src/Toolkit/Loggable/ResultDecorator.php
@@ -16,6 +16,7 @@ use CloudCreativity\Modules\Contracts\Toolkit\Loggable\ContextProvider;
 use CloudCreativity\Modules\Contracts\Toolkit\Loggable\Contextual;
 use CloudCreativity\Modules\Contracts\Toolkit\Result\Error;
 use CloudCreativity\Modules\Contracts\Toolkit\Result\Result;
+
 use function CloudCreativity\Modules\Toolkit\enum_string;
 
 final readonly class ResultDecorator implements ContextProvider