+++ tests/Unit/Toolkit/Loggable/ResultDecoratorTest.php
@@ -26,6 +26,7 @@ use CloudCreativity\Modules\Toolkit\Result\Result;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use UnitEnum;
+
 use function CloudCreativity\Modules\Toolkit\enum_string;
 
 /**